fix(assistant): warm both guardian-delivery cache keys at every persona warm site
Moves warmGuardianBindings into the guardian-delivery reader and uses it
at the four call sites that warmed only the vellum channel key (sweep
job, v2 router, heartbeat, retrospective job). Their sync persona
resolution falls back to peekAnyGuardian, which reads the separate
unfiltered cache key, so a guardian bound only on a non-vellum channel
still landed on users/default.md. Addresses the Codex P2 on #40416.
