+++ assistant/src/contacts/__tests__/guardian-delivery-reader.test.ts
@@ -44,6 +44,7 @@ import {
   getGuardianDeliveryFresh,
   guardianForChannel,
   invalidateGuardianDeliveryCache,
+  warmGuardianBindings,
 } from "../guardian-delivery-reader.js";
 
 const METHOD = "resolve_guardian_delivery";
@@ -297,5 +298,20 @@ describe("selectors", () => {
   test("anyGuardian returns the first overall", () => {
     expect(anyGuardian([emailGuardian, telegramGuardian])).toBe(emailGuardian);
     expect(anyGuardian([])).toBeUndefined();
+  });
+});
+
+describe("warmGuardianBindings", () => {
+  test("warms both the vellum-channel key and the unfiltered fallback key", async () => {
+    const keys: string[] = [];
+    await warmGuardianBindings(async (input) => {
+      keys.push(input?.channelTypes?.join(",") ?? "ALL");
+      return null;
+    });
+    // Both keys the sync persona resolvers read — peekGuardianForChannel
+    // ("vellum") and the peekAnyGuardian() fallback — must be warmed;
+    // warming only the channel key leaves the any-channel fallback cold for
+    // guardians bound on non-vellum channels.
+    expect(keys.sort()).toEqual(["ALL", "vellum"]);
   });
 });
