+++ assistant/src/plugins/defaults/memory/v2/router.ts
@@ -41,7 +41,7 @@ import {
 import { z } from "zod";
 
 import type { AssistantConfig } from "../../../../config/types.js";
-import { getGuardianDelivery } from "../../../../contacts/guardian-delivery-reader.js";
+import { warmGuardianBindings } from "../../../../contacts/guardian-delivery-reader.js";
 import {
   cachedTextBlock,
   extractToolUse,
@@ -394,11 +394,12 @@ async function runRouterBatch(
     provider,
   } = params;
 
-  // Warm the vellum guardian-delivery cache so resolveUserName's sync
-  // guardian-file resolution hits a fresh key instead of falling back to
-  // users/default.md on a cold/TTL-expired cache. Single-flight and
-  // TTL-cached, so per-batch calls collapse to at most one IPC read.
-  await getGuardianDelivery({ channelTypes: ["vellum"] });
+  // Warm both guardian-delivery cache keys (vellum + unfiltered) so
+  // resolveUserName's sync guardian-file resolution, including its
+  // any-channel fallback, hits fresh keys instead of falling back to
+  // users/default.md on a cold/TTL-expired cache. Two local-socket IPC
+  // reads per batch, dwarfed by the batch's provider call.
+  await warmGuardianBindings();
 
   const systemPrompt = resolveRouterPrompt(
     config.memory?.v2?.router?.router_prompt_path ?? null,
